Lots of repairs:
isTeam variable in DeletableAbstractModel class.
Owner String, to represent owner.

Both should be set on the serverside.
Also, changes to List Models to have the backing list be synchronized
using collections.synchronizedSet()
I hope this will prevent race conditions. Supposedly such a List is theadsafe.

Also changes to EventEntityManager
so that the made events get an owner and a UniqueID
that is truly unique, set on ther server.